Draw X, corners, etc. to scratch surface, then blit
We used to directly draw various things on a surface.  We now draw
them on a scratch surface and then blit them to the intended
destination.

We could probably get better performance if we cached some of the
things we draw this way (e.g., had a surface with an empty
titlebar, etc.).