Completely rework Window->surf
The surf is now only the client area; the chrome is drawn completely
independently.  Additionally, it's now totally optional.  If you
create the window with WIN_F_AUTOSURF, you get a surface, and it's
adapted to the window size automatically.  Otherwise, it's up to
you.  If there *is* a surf, it gets drawn for you.  If it's not the
right size, borders are filled in with the window background color
(which is also used to fill in empty areas when resizing the surface
if it's in automatic mode).

Custom memory management for the w->surf pixels is gone, since you
can now take total control of the surf if you want something more
interesting than the default (so no pix_alloc/pix_free anymore).

There's a new on_resized() event which fires after a resize (as
opposed to the on_resize which fires continuously during resize).

This also may have cleaned up some state issues when creating a
window during various modes (e.g., resizing).