Expose more of event loop stuff
The expectation was that you would probably write your own event loop
(or integrate lux into something that already had an event loop), so
the built-in one was really only meant as a toy.

We now expose more pieces of it to make it more useful to other
programs.  In particular, this may be nice when using lux from
Python, which is a new use case.